+++ client/src/Tools/_framework/Paths/ActivityViewer.tsx
@@ -267,7 +267,6 @@ export function ActivityViewer() {
         mode={mode}
         contentId={contentId}
         headerHeight="140px"
-        fetcher={fetcher}
       />
     );
   } else {
