+++ client/src/Tools/_framework/ToolPanels/DoenetMLHistoryModals/CreateRevisionModal.tsx
@@ -19,12 +19,12 @@ import { FetcherWithComponents } from "react-router";
 import { ContentRevision } from "../../../../_utils/types";
 import axios from "axios";
 
-export async function takeSnapshotModelActions({
+export async function createRevisionModelActions({
   formObj,
 }: {
   [k: string]: any;
 }) {
-  if (formObj?._action === "create snapshot") {
+  if (formObj?._action === "create revision") {
     const {
       data: { createdNew },
     } = await axios.post("/api/updateContent/createContentRevision", {
@@ -38,7 +38,7 @@ export async function takeSnapshotModelActions({
   return null;
 }
 
-export function TakeSnapshotModel({
+export function CreateRevisionModel({
   isOpen,
   onClose,
   revisions,
@@ -58,7 +58,7 @@ export function TakeSnapshotModel({
   const [revisionName, setRevisionName] = useState("");
   const [note, setNote] = useState("");
   const [updateRevision, setUpdateRevision] = useState(false);
-  const [snapshotMessage, setSnapshotMessage] = useState("");
+  const [revisionMessage, setRevisionMessage] = useState("");
   const [statusStyleIdx, setStatusStyleIdx] = useState(0);
 
   useEffect(() => {
@@ -67,9 +67,9 @@ export function TakeSnapshotModel({
         setStatusStyleIdx((x) => x + 1);
         setUpdateRevision(true);
         if (fetcher.data.createdNew) {
-          setSnapshotMessage("Created new snapshot (i.e., revision)");
+          setRevisionMessage("Created new revision");
         } else {
-          setSnapshotMessage("Updated snapshot (i.e., revision)");
+          setRevisionMessage("Updated revision");
         }
       }
     }
@@ -78,7 +78,7 @@ export function TakeSnapshotModel({
   useEffect(() => {
     async function checkIfAtLastRevision() {
       // Check to see if last revision has the same cid as the current activity.
-      // If so, preload the take snapshot modal with information about that revision,
+      // If so, preload the modal with information about that revision,
       // as it will be modifying that revision.
 
       if (revisions.length > 0 && atLastRevision) {
@@ -101,7 +101,7 @@ export function TakeSnapshotModel({
 
   useEffect(() => {
     if (isOpen) {
-      setSnapshotMessage("");
+      setRevisionMessage("");
     }
   }, [isOpen]);
 
@@ -116,26 +116,26 @@ export function TakeSnapshotModel({
       <ModalContent>
         <ModalHeader textAlign="center">
           {updateRevision
-            ? "Update info on current snapshot"
-            : "Take snapshot of current activity"}
+            ? "Update info on current revision"
+            : "Create a revision of current activity"}
         </ModalHeader>
         <ModalBody>
           <Text>
             {updateRevision
-              ? "Snapshot of the current state of the activity has already been taken. You can update the description of the revision."
-              : "Take snapshot of the activity to keep a record (a saved revision) of its current state."}
+              ? `${!revisionMessage ? "A revision of current state of the activity already exists." : ""} You can update the description of the revision.`
+              : "Create a revision of the activity to keep a record of its current state."}
           </Text>
 
-          {snapshotMessage !== "" ? (
+          {revisionMessage !== "" ? (
             <Box
               data-test="Status message"
               border="solid 1px lightgray"
               borderRadius="5px"
               padding="5px 10px"
               marginTop="10px"
-              backgroundColor={["orange.100", "orange.200"][statusStyleIdx % 2]}
+              backgroundColor={["green.100", "green.200"][statusStyleIdx % 2]}
             >
-              {snapshotMessage}
+              {revisionMessage}
             </Box>
           ) : null}
 
@@ -160,13 +160,13 @@ export function TakeSnapshotModel({
 
         <ModalFooter>
           <Button
-            data-test="Save Snapshot"
+            data-test="Create revision"
             marginRight="4px"
             isDisabled={revisionName === ""}
             onClick={() => {
               fetcher.submit(
                 {
-                  _action: "create snapshot",
+                  _action: "create revision",
                   contentId,
                   revisionName,
                   note,
