+++ client/src/index.tsx
@@ -95,6 +95,11 @@ import {
   action as doenetMLComparisonAction,
 } from "./Tools/_framework/Paths/DoenetMLComparison";
 import {
+  DoenetMLHistory,
+  loader as doenetMLHistoryLoader,
+  action as doenetMLHistoryAction,
+} from "./Tools/_framework/Paths/DoenetMLHistory";
+import {
   CodeViewer,
   loader as codeViewerLoader,
 } from "./Tools/_framework/Paths/CodeViewer";
@@ -302,6 +307,13 @@ const router = createBrowserRouter([
         loader: doenetMLComparisonLoader,
         action: doenetMLComparisonAction,
         element: <DoenetMLComparison />,
+        errorElement: <ErrorPage />,
+      },
+      {
+        path: "activityHistory/:contentId",
+        loader: doenetMLHistoryLoader,
+        action: doenetMLHistoryAction,
+        element: <DoenetMLHistory />,
         errorElement: <ErrorPage />,
       },
       {
