+++ server/src/routes/remixRoutes.ts
@@ -14,6 +14,7 @@ import {
   updateOriginContentToRemixSchema,
   updateRemixedContentToOriginSchema,
 } from "../schemas/remixSchema";
+import { requireLoggedIn } from "../middleware/validationMiddleware";
 
 export const remixRouter = express.Router();
 
@@ -29,6 +30,7 @@ remixRouter.get(
 
 remixRouter.post(
   "/updateRemixedContentToOrigin",
+  requireLoggedIn,
   queryLoggedIn(
     updateRemixedContentToOrigin,
     updateRemixedContentToOriginSchema,
@@ -37,5 +39,6 @@ remixRouter.post(
 
 remixRouter.post(
   "/updateOriginContentToRemix",
+  requireLoggedIn,
   queryLoggedIn(updateOriginContentToRemix, updateOriginContentToRemixSchema),
 );
